requirements.txt
Django~=3.2.4
requests~=2.25.1
djangorestframework~=3.12.4
django-filter~=2.4.0